Add support for null values, validate on SetValue
When calling SetValue, an empty int array represents null, regardless of the
type of the value. This is in line with Victron's convention on null values.

Here, we also improve handling and converting values when calling SetValue.
Most notably, in the past, we stored a value presented regardless of its type,
now we check the type and rather fail 'SetValue' in case the type does not
match or cannot be converted.

Squashed commit of the following:

commit b74df6770dda9d0ec6ee5af6b8b5eb991d657343
Author: Chris Oloff <chris@uber5.com>
Date:   Thu Feb 13 13:21:46 2025 +0200

    SetValue, improve test coverage

commit 1f8fd9f3be1925659e644c72d166d1242d14700d
Author: Chris Oloff <chris@uber5.com>
Date:   Thu Feb 13 11:56:54 2025 +0200

    SetValue, validate type and min/max constraints

commit 5580cbe569ad978a8e49357dd67dbac64838ae7f
Author: Chris Oloff <chris@uber5.com>
Date:   Thu Feb 13 07:49:48 2025 +0200

    null value represented as empty int array now roughly working, requires cleanup and testing